requirements.txt
appdirs==1.4.4
attrs==19.3.0  
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.4
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.0
django-crispy-forms==1.9.1
django-profanity-check==0.1.0
django-profanity-filter==0.2.1
django-recaptcha==2.0.6
docutils==0.16
entrypoints==0.3
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
inflection==0.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.1.0
itermate==1.0.2
jedi==0.16.0
Jinja2==2.11.1
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.1
pandocfilters==1.4.2
parsel==1.6.0
parso==0.6.2
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2020.6.2
profanity-check==1.0.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyrsistent==0.16.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
queuelib==1.5.0
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
Scrapy==2.2.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
somewhere==0.0.1
soupsieve==2.0.1
SQLAlchemy==1.3.13
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.9
virtualenv==16.7.7
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.1.9
webencodings==0.5.1
where==1.0.2
widgetsnbextension==3.5.1
zipp==3.1.0
zope.interface==5.1.0